requirements.txt
aiohttp==3.5.4
certifi==2019.9.11
chardet==3.0.4
discord.py==1.2.4
gmplot==1.2.0
idna==2.8
requests==2.22.0
urllib3==1.25.6
websockets==6.0
youtube-dl==2019.10.29
PyNaCl==1.3.0
wikipedia==1.4.0
